third_party/blink/web_tests/wpt_internal/svg/struct/systemLanguage-reflect-user-language-in-image.html

<!DOCTYPE HTML>
<title>SVG systemLanguage works in image contexts</title>
<link rel="help" href="https://www.w3.org/TR/SVG2/struct.html#ConditionalProcessingSystemLanguageAttribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<script>

promise_test(async () => {
  internals.settings.setAcceptLanguages("en");

  const canvas = new OffscreenCanvas(100, 100);
  const ctx = canvas.getContext('2d');

  // Load an SVG-as-image with a <switch> using systemLanguage.
  // If systemLanguage="en" matches, a green rect renders.
  // If it doesn't match (the bug), the red fallback renders.
  const img = new Image();
  const loaded = new Promise(resolve => img.onload = resolve);
  img.src = "data:image/svg+xml," +
    "<svg xmlns='http://www.w3.org/2000/svg' width='100' height='100'>" +
    "<switch>" +
    "<rect systemLanguage='en' width='100' height='100' fill='green'/>" +
    "<rect width='100' height='100' fill='red'/>" +
    "</switch>" +
    "</svg>";
  await loaded;

  ctx.drawImage(img, 0, 0);
  _assertPixel(canvas, 50, 50, 0, 128, 0, 255);
}, "systemLanguage conditional processing works inside SVG loaded as an image");

</script>
